ai_ops_assistant/requirements.txt
# LLM Provider
google-generativeai>=0.3.0

# API Integrations
requests>=2.31.0
PyGithub>=2.1.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Web Framework (FastAPI)
fastapi>=0.104.1
uvicorn>=0.24.0

# Optional: CLI enhancement
rich>=13.7.0
